--- link_prediction/karate.csv
@@ -1,78 +1,0 @@
-1,32,{}
-1,2,{}
-1,3,{}
-1,4,{}
-1,5,{}
-1,6,{}
-1,7,{}
-1,8,{}
-1,9,{}
-1,11,{}
-1,12,{}
-1,13,{}
-1,14,{}
-1,18,{}
-1,20,{}
-1,22,{}
-2,3,{}
-2,4,{}
-2,8,{}
-2,14,{}
-2,18,{}
-2,20,{}
-2,22,{}
-2,31,{}
-3,4,{}
-3,33,{}
-3,8,{}
-3,9,{}
-3,10,{}
-3,14,{}
-3,28,{}
-3,29,{}
-4,8,{}
-4,13,{}
-4,14,{}
-5,11,{}
-5,7,{}
-6,7,{}
-6,11,{}
-6,17,{}
-7,17,{}
-9,31,{}
-9,34,{}
-9,33,{}
-10,34,{}
-14,34,{}
-15,33,{}
-15,34,{}
-16,33,{}
-16,34,{}
-19,33,{}
-19,34,{}
-20,34,{}
-21,33,{}
-21,34,{}
-23,33,{}
-23,34,{}
-24,33,{}
-24,26,{}
-24,28,{}
-24,34,{}
-24,30,{}
-25,32,{}
-25,26,{}
-25,28,{}
-26,32,{}
-27,34,{}
-27,30,{}
-28,34,{}
-29,32,{}
-29,34,{}
-30,33,{}
-30,34,{}
-31,34,{}
-31,33,{}
-32,33,{}
-32,34,{}
-33,34,{}
