+++ main.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from OpenSNA import Graph
 
-edge_list = 'edges.csv'
+edge_list = '../../data/edges.csv'
 
 g = Graph()
 g.load_graph(edge_list, False)
